moved widget to global component
