+++ lib/features/component/glabal_shared/stark_amount.dart
