+++ lib/features/feature.dart
@@ -65,7 +65,8 @@ part 'wallet_screen/widget/waltet_screen_body.dart';
 part 'wallet_screen/widget/wallet_screen_tablet_mode.dart';
 part 'wallet_screen/widget/contract_address.dart';
 part 'wallet_screen/widget/recent_transaction.dart';
-part 'wallet_screen/widget/stark_amount.dart';
 part 'wallet_screen/widget/transaction_item.dart';
 part 'wallet_screen/widget/wallet_body_padding.dart';
-part 'wallet_screen/widget/tablet_header.dart';
+
+part 'component/glabal_shared/stark_amount.dart';
+part 'component/glabal_shared/tablet_header.dart';
